+++ media_kit/lib/src/models/player_error.dart
@@ -4,7 +4,14 @@
 /// All rights reserved.
 /// Use of this source code is governed by MIT license that can be found in the LICENSE file.
 
+/// {@template player_error}
+///
+/// PlayerError
+/// -----------
+///
 /// An error experienced by [Player] during playback.
+///
+/// {@endtemplate}
 class PlayerError {
   /// Error code.
   final int code;
@@ -12,11 +19,12 @@ class PlayerError {
   /// Error message.
   final String message;
 
+  /// {@macro player_error}
   const PlayerError(
     this.code,
     this.message,
   );
 
   @override
-  String toString() => 'PlayerError(code: $code, message: $message)';
+  String toString() => 'PlayerError($code, $message)';
 }