+++ media_kit_video/CHANGELOG.md
@@ -6,6 +6,7 @@
 - iOS support:
   - Hardware: MPV_RENDER_API_TYPE_OPENGL + pixel buffer
   - Software: MPV_RENDER_API_TYPE_SW + pixel buffer
+- fix: use `TextureRegistrar::UnregisterTexture` release callback to free texture resources on Windows.
 - feat: `aspectRatio` parameter for `Video` widget.
 
 ## 0.0.1