+++ media_kit_video/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+- fix: add `.framework` & `.xcframework` for all libs
+
 ## 0.0.2
 
 - macOS support: